+++ CHANGELOG.md
@@ -7,6 +7,7 @@ HEAD
 * Add `ElementProxy::memoryUsage()`
 * Add `MemberProxy::memoryUsage()` (issue #1730)
 * Add implicit conversion from `JsonDocument` to `JsonVariant`
+* Fix comparisons operators with `const JsonDocument&`
 
 v6.19.3 (2022-03-08)
 -------